gtaxl: Revert "workaround for offline charging"
This reverts commit 667ef4c515ec8256e84abb7eb7f04188899527d9.

This workaround caused kernel panics due to "sysmmu page read fault[s]"
that, of course, caused full reboots when trying to charge offline, as
intended. It always assumes that the device has enough power to boot,
making it a bad workaround. Now that the property which enabled booting
straight from offline charging without a full reboot, which caused issues
with bootlooping when trying to bootup from offline charging that this
workaround was just trying to get around the inconveniences of, is removed
thereby disabling that feature, no workaround is necessary.

Change-Id: I45af1d120d250332d6fa74022324a06a2035767f