gtaxl: Symlink vendor/lib[64]/egl/libGLES_mali.so to vendor/lib[64]/vulkan.exynos5.so, rather than vendor/lib[64]/hw/vulkan.exynos5.so, and stop copying
On 10, we were copying the libGLES_mali.so blobs to
vendor/lib[64]/hw/vulkan.exynos5.so rather than symlinking as otherwise,
Vulkan wouldn't work.

Now, on 11, that no longer works. The following error occurs when trying to
use Vulkan:
04-02 16:13:25.177 10089 10089 E vndksupport: Could not load vulkan.exynos5.so from default namespace: dlopen failed: library "vulkan.exynos5.so" not found.

Moving vulkan.exynos5.so to vendor/lib[64]/ fixes Vulkan on 11. Additionally,
it also looks like it also works with the libGLES_mali.so blobs symlinked to
vulkan.exynos5.so in the new locations, so we can now stop copying them.

Change-Id: I0405a9d86413a79779eb80b35602b4c5cb193c3d