gtaxlwifi: Add gpsd to class hal
With the previous setup where gpsd is started after boot, now, with 11, that
no longer works when SELinux is enforcing. There appeared to be some type of
issues with starting gpsd and the gnss HAL, and there were no relevant SELinux
denials reported in logs. The gnss HAL would need to be restarted by running
`stop vendor.gnss_service` and `start vendor.gnss_service` to get GPS to work.

With gpsd in class hal, those issues have disappeared, and GPS works. I'm not
sure if this is technically right, but I don't have any other solution, other
than the workaround of restarting the gnss HAL after booting, and it works
flawlessly!

Change-Id: Ib2a02823dbc9586f2c50ffe3ee60f0c234389109