gtaxl-common: Declare support for pro audio
With the OSS audio HAL, and with reduced playback and low-latency
capture period sizes, audio latency is now much lower.

According to tests with the Superpowered audio latency test app [1],
with very rough tests with the Dr. Rick O'Rang loopback app [2][3]
with just my wired headphones and the built-in microphone confirming,
on my tablet, round-trip audio latency is now around 14-16 ms.
Therefore, seeing as all of the other requirements for pro audio
support to be reported seem to be satisfied as well, it's now doable.

 [1] https://superpowered.com/latency
 [2] https://source.android.com/devices/audio/latency/measure#measuringRoundTrip
 [3] https://github.com/gkasten/drrickorang/tree/master/LoopbackApp

Change-Id: I8ba63da90e7309afb3c06cb3ce80c1848c6f4f91