gtaxl-common: Switch to Bluetooth firmware from P585MUBU2CUF2
This fixes call audio over Bluetooth. With the previous firmware, it
simply wouldn't work at all.

Thankfully, this isn't at all a downgrade of the firmware, from when
I previously updated it from a stock firmware for another device.

With this switch, the rampatch_tlv_tf_1.1.tlv file remains the same as
from the previous update. The other firmware file that is changed with
this switch, nvm_tlv_tf_1.1.bin, has turned out to be specific to these
devices.

Change-Id: Iec097094457d80d662fe13abb6fdb3fdb75928c7