gtaxl-common: audio: mixer_paths: Make the dev-codec-out-disable path useful
The dev-codec-out-disable path is triggered for calls over Bluetooth,
and it should be explicitly disabling the speaker and headphone outputs,
which it was not.

Previously, without it actually explicitly disabling the speaker and
headphone outputs, while in a call, audio would undesirably be output
from either the speakers or headphone output as well as through
Bluetooth.

Change-Id: Id540558ef6f613f5fd90763e70c8c09aeaa48561