gtaxl-common: init: Explicitly disable secure ADB in recovery
For that, this sets the ro.adb.secure.recovery property to 0. It appears
it is supposed to be set after selecting an "Enable ADB" option in
Lineage recovery, but for some reason, something goes wrong and ADB
stays stuck in an unauthorized state, and setting it explicitly works
around that problem.

On a related note, ADB sideload nevertheless doesn't work seemingly due
to the old FunctionFS implementation in the kernel.

Change-Id: Ia4ec0d7bf5664460e284dac4d4f505fcf5919962