gtaxl-common: audio: Revert audio setup back to commit 8633a5c8e
This reverts commits:
 - 7b0a0db122c ("gtaxl-common: audio: Bring mixer_paths files in line
 with stock")
 - 4c9458668c3 ("gtaxl-common: Pull in custom libaudioroute")
 - 649d656280a ("gtaxl-common: audio_hal: Revert "Forcefully reset and
 update audio route paths"")
 - c2d8553f7a7 ("gtaxl-common: audio: mixer_paths_rev07: Revert issue
 fix attempt")
 - c4bc2b2f895 ("gtaxl-common: audio_hal: Forcefully reset and update
 audio route paths")

Unfortunately, all my attempts to fix audio issues on HW revision 7
gtaxllte tablets have ended up apparently not being successful. The
state of the audio setup at commit 8633a5c8e was the cleanest, and
excluding one issue that I'll fix in the next commit, it worked
perfectly fine on non-revision 7 tablets, so this reverts back to it.

Change-Id: I22b379d6de7c7d49a48fceb15f25df60d5ea7051